Added UniqueFilename
Tries to return a UniqueFilename...

Written for use with Temp files in a threaded situation.   But also provided a more directory friendly format in case 255 is an issue.  Friendly version is NOT threadsafe.   I need to start using this generically...  (instead of stuff like temp.txt, which runs the risk of two different calls overwriting each other)